fixed env and database init
- fixed env and database init

- fixed type errors and styles
