+++ .env.example
@@ -1,3 +1,6 @@
 EXPO_PUBLIC_API_URL="http://localhost:8081"
 EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY="Your Expo Public Clerk Publishable Key Here"
+# keep Clerk secret key secure: do not expose it to client-side code
+CLERK_SECRET_KEY=sk_test_dqIE0M1p1CbgwQOD63f9DTMAwo53u93hOjO6xRMjtd
+# keep Neon credentials secure: do not expose them to client-side code
 DATABASE_URL="postgresql://<USERNAME>:<PASSWORD>@<HOST>/<DATABASE>?sslmode=require"
