+++ src/db/index.ts
@@ -1,11 +1,6 @@
-import { drizzle } from "drizzle-orm/neon-serverless";
-import { neon, neonConfig } from "@neondatabase/serverless";
-import { env } from "../env";
-
-neonConfig.useSecureWebSocket = true;
-neonConfig.fetchConnectionCache = true;
-
+import { neon } from "@neondatabase/serverless";
+import { drizzle } from "drizzle-orm/neon-http";
 const databaseUrl = process.env.DATABASE_URL;
 if (!databaseUrl) throw new Error("DATABASE_URL is not set");
-
-export const db = drizzle(databaseUrl);
+const sql = neon(databaseUrl);
+export const db = drizzle({ client: sql });
